backend/requirements.txt
flask
flask_cors
gevent
mysql-connector-python